refactor: Move lib and db folders to infrastructure layer
- Move src/lib/ to src/infrastructure/lib/
- Move src/db/ to src/infrastructure/persistence/drizzle-db/
- Update all import paths from @/lib to @/infrastructure/lib
- Update drizzle.config.ts schema path
- Update src/index.ts db import path
