+++ drizzle.config.ts
@@ -2,7 +2,7 @@ import type { Config } from 'drizzle-kit';
 import { env } from './src/env';
 
 export default {
-  schema: './src/db/schema.ts',
+  schema: './src/infrastructure/persistence/drizzle-db/schema.ts',
   out: './drizzle',
   dialect: 'postgresql',
   dbCredentials: {
