+++ src/infrastructure/lib/db.ts
@@ -1,7 +1,7 @@
 import { drizzle } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
-import * as schema from '../db/schema';
-import { env } from '../env';
+import * as schema from '../persistence/drizzle-db/schema';
+import { env } from '../../env';
 
 const client = postgres(env.DATABASE_URL);
 export const db = drizzle(client, { schema });
