+++ src/presentation/http/github/github.handlers.ts
@@ -1,4 +1,4 @@
-import type { AppContext } from '@/lib/router';
+import type { AppContext } from '@/infrastructure/lib/router';
 import * as HttpStatusCodes from 'stoker/http-status-codes';
 import type { z } from 'zod';
 import type {
