+++ src/presentation/http/github/github.index.ts
@@ -1,4 +1,4 @@
-import { createRouter } from '@/lib/router';
+import { createRouter } from '@/infrastructure/lib/router';
 import * as routes from './github.routes';
 import * as handlers from './github.handlers';
