+++ src/presentation/http/github/github.routes.ts
@@ -1,4 +1,7 @@
-import { InternalServerErrorSchema, NotFoundErrorSchema } from '@/lib/error';
+import {
+  InternalServerErrorSchema,
+  NotFoundErrorSchema,
+} from '@/infrastructure/lib/error';
 import { createRoute, z } from '@hono/zod-openapi';
 import * as HttpStatusCodes from 'stoker/http-status-codes';
 import { jsonContent, jsonContentRequired } from 'stoker/openapi/helpers';
