+++ src/presentation/http/reminder/reminder.index.ts
@@ -1,4 +1,4 @@
-import { createRouter } from '@/lib/router';
+import { createRouter } from '@/infrastructure/lib/router';
 import * as routes from './reminder.routes';
 import * as handlers from './reminder.handlers';
